requirements.txt
autopep8==2.0.2
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
google-auth==2.16.3
google-auth-oauthlib==1.0.0
gspread==5.7.2
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyparsing==3.0.9
python-dotenv==1.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tomli==2.0.1
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
